chore(refactoring): Move Store and delete TestDocumentFactory
