+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/document/store/CamundaDocumentStoreImpl.java
@@ -17,7 +17,6 @@
 package io.camunda.connector.runtime.core.document.store;
 
 import io.camunda.client.CamundaClient;
-import io.camunda.connector.api.document.CamundaDocumentStore;
 import io.camunda.connector.api.document.DocumentCreationRequest;
 import io.camunda.connector.api.document.DocumentLinkParameters;
 import io.camunda.connector.api.document.DocumentReference.CamundaDocumentReference;
