+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/document/store/InMemoryDocumentStore.java
@@ -16,7 +16,6 @@
  */
 package io.camunda.connector.runtime.core.document.store;
 
-import io.camunda.connector.api.document.CamundaDocumentStore;
 import io.camunda.connector.api.document.DocumentCreationRequest;
 import io.camunda.connector.api.document.DocumentLinkParameters;
 import io.camunda.connector.api.document.DocumentMetadata;
